Counter added

+++ PrimeCounter/src/main/java/com.count.prime.counter/MapActor.java
@@ -11,7 +11,7 @@ public class MapActor extends UntypedActor{
     public void onReceive(Object message) throws Exception {
        if(message instanceof Size){
            Size counter=(Size) message;
-           for(Long i=0l;i<counter.getSize();i++){
+           for(int i=0;i<counter.getSize();i++){
               getSender().tell(i);
            }
        }else{
