+++ PrimeCounter/src/main/java/com.count.prime.counter/PrimeVerifier.java
@@ -7,8 +7,8 @@ public class PrimeVerifier {
     //No state
 
     //checks whether an int is prime or not.
-    public static boolean isPrime(long n) {
-        for(long i=2;i<n;i++) {
+    public static boolean isPrime(int n) {
+        for(int i=2;i<n;i++) {
             if(n%i==0)
                 return false;
         }
