+++ PrimeCounter/src/main/java/messages/Number.java
@@ -4,10 +4,10 @@ package messages;
  * Created by imran on 1/7/17.
  */
 public class Number {
-    private Long number;
+    private Integer number;
     private Boolean prime;
 
-    public Number(Long number, Boolean prime) {
+    public Number(Integer number, Boolean prime) {
         this.number = number;
         this.prime = prime;
     }
@@ -16,7 +16,7 @@ public class Number {
         return prime;
     }
 
-    public Long getNumber() {
+    public Integer getNumber() {
         return number;
     }
