+++ PrimeCounter/src/main/java/messages/Size.java
@@ -4,14 +4,14 @@ package messages;
  * Created by imran on 1/7/17.
  */
 public class Size {
-    private Long size;
+    private Integer size;
 
-    public Size(Long size){
+    public Size(Integer size){
         this.size=size;
     }
 
 
-    public Long getSize() {
+    public Integer getSize() {
         return size;
     }
 }
